Add LICENSE (EPL-1.0)
This code base is licensed under EPL-1.0 because it embeds other code
that is licensed under EPL-1.0. I would rather use MIT, but since this
is not a library, contacting the author of the libraries I'm embedding
does not seem worth the hassle. Appologies to anyone who just wants to
grab code of this repo and use it. I will attempt to turn anything
useful and generalizeable in this repo into an MIT license library as I
have time.